chore: 스토리북 preview 설정 수정
- theme. global 스타일이 적용되도록 설정
